coancestries calculate

Given genetic data it provides coancestry for between and within populations

Usage

coancestries_calculate(genetic_data_parents, genetic_data_F1, datatype = "dosage", number_of_populations = NA, population_individual_id = NA, pedigree = NA, usepedigree = FALSE, BiAllelic = TRUE)

Arguments

genetic_data_parents

A data frame or file name containing genetic data for the parental generation. The format should match the specified datatype (e.g., dosage, VCF, or dat which is FSTAT format).

genetic_data_F1

A data frame or file name containing genetic data for the F1 generation. The format should match the specified datatype (e.g., dosage, VCF, or dat which is FSTAT format).

datatype

A character string specifying the type of genetic data: "dosage" (default), "dat" (dataframe), "FSTAT" format), or "vcf" (variant call format).

number_of_populations

(Optional) A scalar indicating the number of populations.

population_individual_id

(Optional) A data frame mapping individuals to their respective populations. First parent ids, then F1 ids. If not provided, a balanced design is assumed.

pedigree

(Optional) A data frame containing pedigree information with at least three columns: individual ID, dam ID, and sire ID.

usepedigree

Logical; if TRUE, kinship calculations will be based on the pedigree rather than genetic data. Defaults to FALSE.

BiAllelic

Logical; if TRUE, assumes biallelic markers for certain genetic data formats. Defaults to TRUE.

Value

A list containing:

The.M

A kinship-based relatedness matrix adjusted for F1 individuals.

Theta.P

An adjusted population coancestry matrix.

Details

This function calculates kinship and coancestry measures from genetic data or pedigree-based information, adjusting for population structure and F1 generation characteristics.

References

Goudet & Weir (2023)

Author

Isabela do O isabela.doo@unil.ch