lava: Log ratios of ancestral variances

Given coancestry matrices for between and within populations and a trait data frame, this function estimates the log ratio of ancestral variances using a Bayesian mixed-effects model.

Usage

lava(Theta.P, The.M, trait_dataframe, column_individual = "id", column_trait = "trait", 
            column_population = "population", formula_covariates = NULL, ...)

Arguments

Theta.P

A square matrix representing the coancestry matrix between populations.

The.M

A square matrix representing the kinship-based relatedness matrix for individuals.

trait_dataframe

A data frame containing individual IDs and trait values. The first column should be individual IDs, and the second column should be trait values.

column_individual

The name of the column containing individual IDs. Default is "id".

column_trait

The name of the column containing trait values. Default is "trait".

column_population

The name of the column containing population IDs. Default is "population".

formula_covariates

A character string specifying additional covariates to include in the model. For example, "age + sex" would add age and sex as fixed effects. Default is NULL (no covariates).

...

Additional arguments passed to the brms function.

Value

A lava type object containing:

posterior_samples

A list with posterior samples of variance components and residuals.

log_ratio

A list with the p-value of the log ratios, the mean of the log ratio of between- and within-population variance, and confidence intervals.

model

as an attribute.

Details

This function standardizes trait data, constructs a Bayesian mixed-effects model using brms, and estimates ancestral variances. The function assumes no crosses between populations and analyzes one trait at a time.

References

Goudet & Weir (2023)

do O et al (2025)

Author

Isabela do O isabela.doo@unil.ch