lava.options {lava}	R Documentation

Set global options for lava

Description

Extract and set global parameters of lava. In particular optimization parameters for the estimate function.

Usage

lava.options(...)

Arguments

...
Arguments

Details

param: 'relative' (factor loading and variance of one endogenous variables in each measurement model are fixed to one), 'absolute' (mean and variance of latent variables are set to 0 and 1, respectively), 'hybrid' (intercept of latent variables is fixed to 0, and factor loading of at least one endogenous variable in each measurement model is fixed to 1), 'none' (no constraints are added)

layout: One of 'dot','fdp','circo','twopi','neato','osage'

messages: Set to 0 to disable various output messages

...

see control parameter of the estimate function.

Value

list of parameters

Author(s)

Klaus K. Holst

Examples

## Not run: 
lava.options(iter.max=100,messages=0)

## End(Not run)

[Package lava version 1.6.5 Index]