org {lava}	R Documentation

Convert object to ascii suitable for org-mode

Description

Convert object to ascii suitable for org-mode

Usage

org(x, ..., ncol, include.rownames = TRUE, include.colnames = TRUE,
  header = TRUE, frame = "topbot", rownames = NULL, colnames = NULL,
  type = "org", tab = FALSE, margins = TRUE, print = TRUE, html, latex)

Arguments

x
R object

...
additional arguments to lower level functions

ncol
If x is a vector and ncol is given as argument, the resulting output will be a matrix with ncol columns

include.rownames
If FALSE row names are removed

include.colnames
If FALSE column names are removed

header
If TRUE the header is included

frame
Frame argument (see ascii)

rownames
Optional vector of row names

colnames
Optional vector of column names

type
Type argument (see ascii)

tab
Tabulate?

margins
Add margins to table?

print
print or return result

html
HTML prefix (added to ATTR_HTML)

latex
LaTeX prefix (added to ATTR_LaTeX)

Author(s)

Klaus K. Holst

[Package lava version 1.4.4 Index]