partialcor {lava}	R Documentation

Calculate partial correlations

Description

Calculate partial correlation coefficients and confidence limits via Fishers z-transform

Usage

partialcor(formula, data, level = 0.95, ...)

Arguments

formula
formula speciying the covariates and optionally the outcomes to calculate partial correlation for

data
data.frame

level
Level of confidence limits

...
Additional arguments to lower level functions

Value

A coefficient matrix

Author(s)

Klaus K. Holst

Examples

m <- lvm(c(y1,y2,y3)~x1+x2)
covariance(m) <- c(y1,y2,y3)~y1+y2+y3
d <- sim(m,500)
partialcor(~x1+x2,d)

[Package lava version 1.6.5 Index]