subset.lvm {lava}	R Documentation

Extract subset of latent variable model

Description

Extract measurement models or user-specified subset of model

Usage

## S3 method for class 'lvm'
subset(x, vars, ...)

Arguments

x
lvm-object.

vars
Character vector or formula specifying variables to include in subset.

...
Additional arguments to be passed to the low level functions

Value

A lvm-object.

Author(s)

Klaus K. Holst

Examples

m <- lvm(c(y1,y2)~x1+x2)
subset(m,~y1+x1)

[Package lava version 1.6.5 Index]