By {lava}	R Documentation

Apply a Function to a Data Frame Split by Factors

Description

Apply a Function to a Data Frame Split by Factors

Usage

By(x, INDICES, FUN, COLUMNS, array = FALSE, ...)

Arguments

x
Data frame

INDICES
Indices (vector or list of indices, vector of column names, or formula of column names)

FUN
A function to be applied to data frame subsets of 'data'.

COLUMNS
(Optional) subset of columns of x to work on

array
if TRUE an array/matrix is always returned

...
Additional arguments to lower-level functions

Details

Simple wrapper of the 'by' function

Author(s)

Klaus K. Holst

Examples

By(datasets::CO2,~Treatment+Type,colMeans,~conc)
By(datasets::CO2,~Treatment+Type,colMeans,~conc+uptake)

[Package lava version 1.6.5 Index]