bootstrap.lvm {lava}	R Documentation

Calculate bootstrap estimates of a lvm object

Description

Draws non-parametric bootstrap samples

Usage

## S3 method for class 'lvm'
bootstrap(x,R=100,data,fun=NULL,control=list(),
                          p, parametric=FALSE, bollenstine=FALSE,
                          constraints=TRUE,sd=FALSE,messages=lava.options()$messages,
                          parallel=lava.options()$parallel,
                          mc.cores=NULL,
                          ...)

## S3 method for class 'lvmfit'
bootstrap(x,R=100,data=model.frame(x),
                             control=list(start=coef(x)),
                             p=coef(x), parametric=FALSE, bollenstine=FALSE,
                             estimator=x$estimator,weights=Weights(x),...)

Arguments

x
lvm-object.

R
Number of bootstrap samples

data
The data to resample from

fun
Optional function of the (bootstrapped) model-fit defining the statistic of interest

control
Options to the optimization routine

p
Parameter vector of the null model for the parametric bootstrap

parametric
If TRUE a parametric bootstrap is calculated. If FALSE a non-parametric (row-sampling) bootstrap is computed.

bollenstine
Bollen-Stine transformation (non-parametric bootstrap) for bootstrap hypothesis testing.

constraints
Logical indicating whether non-linear parameter constraints should be included in the bootstrap procedure

sd
Logical indicating whether standard error estimates should be included in the bootstrap procedure

messages
Control amount of messages printed

parallel
If TRUE parallel backend will be used

mc.cores
Number of threads (if NULL foreach::foreach will be used, otherwise parallel::mclapply)

...
Additional arguments, e.g. choice of estimator.

estimator
String definining estimator, e.g. 'gaussian' (see estimator)

weights
Optional weights matrix used by estimator

Value

A bootstrap.lvm object.

Author(s)

Klaus K. Holst

See Also

confint.lvmfit

Examples

m <- lvm(y~x)
d <- sim(m,100)
e <- estimate(lvm(y~x), data=d)
 ## Reduce Ex.Timings
B <- bootstrap(e,R=50,parallel=FALSE)
B

[Package lava version 1.6.5 Index]