complik {lava}	R Documentation

Composite Likelihood for probit latent variable models

Description

Estimate parameters in a probit latent variable model via a composite likelihood decomposition.

Usage

complik(x, data, k = 2, type = c("nearest", "all"), pairlist,
  messages = 0, estimator = "normal", ...)

Arguments

x
lvm-object

data
data.frame

k
Size of composite groups

type
Determines number of groups. With type="nearest" (default) only neighboring items will be grouped, e.g. for k=2 (y1,y2),(y2,y3),... With type="all" all combinations of size k are included

pairlist
A list of indices specifying the composite groups. Optional argument which overrides k and type but gives complete flexibility in the specification of the composite likelihood

messages
Control amount of messages printed

estimator
Model (pseudo-likelihood) to use for the pairs/groups

...
Additional arguments parsed on to lower-level functions

Value

An object of class clprobit inheriting methods from lvm

Author(s)

Klaus K. Holst

See Also

estimate

Examples

m <- lvm(c(y1,y2,y3)~b*x+1*u[0],latent=~u)
ordinal(m,K=2) <- ~y1+y2+y3
d <- sim(m,50,seed=1)
e1 <- complik(m,d,control=list(trace=1),type="all")

[Package lava version 1.6.5 Index]