diagtest {lava}	R Documentation

Calculate diagnostic tests for 2x2 table

Description

Calculate prevalence, sensitivity, specificity, and positive and negative predictive values

Usage

diagtest(table, positive = 2, exact = FALSE, p0 = NA,
  confint = c("logit", "arcsin", "pseudoscore", "exact"), ...)

Arguments

table
Table or (matrix/data.frame with two columns)

positive
Switch reference

exact
If TRUE exact binomial proportions CI/test will be used

p0
Optional null hypothesis (test prevalenc, sensitivity, ...)

confint
Type of confidence limits

...
Additional arguments to lower level functions

Details

Table should be in the format with outcome in columns and test in rows. Data.frame should be with test in the first column and outcome in the second column.

Author(s)

Klaus Holst

Examples

M <- as.table(matrix(c(42,12,
                       35,28),ncol=2,byrow=TRUE,
                     dimnames=list(rater=c("no","yes"),gold=c("no","yes"))))
diagtest(M,exact=TRUE)

[Package lava version 1.6.5 Index]