Estimate mixture latent variable model.

Estimate mixture latent variable model

mixture(
  x,
  data,
  k = length(x),
  control = list(),
  vcov = "observed",
  names = FALSE,
  ...
)

Arguments

x
List of lvm objects. If only a single lvm object is given, then a k-mixture of this model is fitted (free parameters varying between mixture components).

data
data.frame

k
Number of mixture components

control
Optimization parameters (see details) #type Type of EM algorithm (standard, classification, stochastic)

vcov
of asymptotic covariance matrix (NULL to omit)

names
If TRUE returns the names of the parameters (for defining starting values)

...
Additional arguments parsed to lower-level functions

Details

Estimate parameters in a mixture of latent variable models via the EM algorithm.

The performance of the EM algorithm can be tuned via the control argument, a list where a subset of the following members can be altered:

start

Optional starting values

nstart

Evaluate nstart different starting values and run the EM-algorithm on the parameters with largest likelihood

tol

Convergence tolerance of the EM-algorithm. The algorithm is stopped when the absolute change in likelihood and parameter (2-norm) between successive iterations is less than tol

iter.max

Maximum number of iterations of the EM-algorithm

gamma

Scale-down (i.e. number between 0 and 1) of the step-size of the Newton-Raphson algorithm in the M-step

trace

Trace information on the EM-algorithm is printed on every traceth iteration

Note that the algorithm can be aborted any time (C-c) and still be saved (via on.exit call).

See also

Examples

# \donttest{
m0 <- lvm(list(y~x+z,x~z))
distribution(m0,~z) <- binomial.lvm()
d <- sim(m0,2000,p=c("y~z"=2,"y~x"=1),seed=1)

## unmeasured confounder example
m <- baptize(lvm(y~x, x~1));
intercept(m,~x+y) <- NA

set.seed(42)
M <- mixture(m,k=2,data=d,control=list(trace=1,tol=1e-6))

#> Squarem-2 
#> Residual:  0.02701133   Extrapolation:  TRUE   Steplength:  1 
#> Residual:  0.03759283   Extrapolation:  TRUE   Steplength:  4 
#> Residual:  0.6706137   Extrapolation:  TRUE   Steplength:  16 
#> Residual:  0.0202431   Extrapolation:  TRUE   Steplength:  2.070703 
#> Residual:  0.001280268   Extrapolation:  TRUE   Steplength:  3.588919 
#> Residual:  0.0003419158   Extrapolation:  TRUE   Steplength:  6.54822 
#> Residual:  0.0001613938   Extrapolation:  TRUE   Steplength:  4.503076 
#> Residual:  0.000107598   Extrapolation:  TRUE   Steplength:  5.973049 
#> Residual:  2.075485e-05   Extrapolation:  TRUE   Steplength:  3.791125 
#> Residual:  2.54089e-06   Extrapolation:  TRUE   Steplength:  4.75328 

summary(M)

#> Cluster 1 (n=1083, Prior=0.5374):
#> --------------------------------------------------
#>                     Estimate Std. Error Z value  Pr(>|z|)
#> Regressions:                                             
#>    y~x               1.04843  0.03919   26.74956   <1e-12
#> Intercepts:                                              
#>    y                 0.04735  0.05471    0.86553 0.3867  
#>    x                 0.06609  0.05609    1.17828 0.2387  
#> Residual Variances:                                      
#>    y                 1.06076  0.06475   16.38310         
#>    x                 1.09065  0.05243   20.80367         
#> 
#> Cluster 2 (n=917, Prior=0.4626):
#> --------------------------------------------------
#>                     Estimate Std. Error Z value  Pr(>|z|)
#> Regressions:                                             
#>    y~x               1.04843  0.03919   26.74956   <1e-12
#> Intercepts:                                              
#>    y                 2.00371  0.07059   28.38673   <1e-12
#>    x                 1.02906  0.05667   18.15727   <1e-12
#> Residual Variances:                                      
#>    y                 1.06076  0.06475   16.38310         
#>    x                 1.09065  0.05243   20.80367         
#> --------------------------------------------------
#> AIC= 13096.52 
#> ||score||^2= 1.349112e-06 

lm(y~x,d)

#> 
#> Call:
#> lm(formula = y ~ x, data = d)
#> 
#> Coefficients:
#> (Intercept)            x  
#>       0.771        1.403  
#> 

estimate(M,"y~x")

#>       Estimate Std.Err   2.5% 97.5%    P-value
#> [y~x]    1.048 0.03811 0.9737 1.123 1.263e-166
#> 
#>  Null Hypothesis: 
#>   [y~x] = 0 

## True slope := 1
# }