path {lava}	R Documentation

Extract pathways in model graph

Description

Extract all possible paths from one variable to another connected component in a latent variable model. In an estimated model the effect size is decomposed into direct, indirect and total effects including approximate standard errors.

Usage

## S3 method for class 'lvm'
 path(object, to = NULL, from, all=FALSE, ...)
## S3 method for class 'lvmfit'
 effects(object, to, from, ...)

Arguments

object
Model object (lvm)

...
Additional arguments to be passed to the low level functions

to
Outcome variable (string). Alternatively a formula specifying response and predictor in which case the argument from is ignored.

from
Response variable (string), not necessarily directly affected by to.

all
If TRUE all simple paths (in undirected graph) is returned on/off.

Value

If object is of class lvmfit a list with the following elements is returned

idx
A list where each element defines a possible pathway via a integer vector indicating the index of the visited nodes.

V
A List of covariance matrices for each path.

coef
A list of parameters estimates for each path

path
A list where each element defines a possible pathway via a character vector naming the visited nodes in order.

edges
Description of 'comp2'

If object is of class lvm only the path element will be returned.

The effects method returns an object of class effects.

Note

For a lvmfit-object the parameters estimates and their corresponding covariance matrix are also returned. The effects-function additionally calculates the total and indirect effects with approximate standard errors

Author(s)

Klaus K. Holst

See Also

children, parents

Examples

m <- lvm(c(y1,y2,y3)~eta)
regression(m) <- y2~x1
latent(m) <- ~eta
regression(m) <- eta~x1+x2
d <- sim(m,500)
e <- estimate(m,d)

path(Model(e),y2~x1)
parents(Model(e), ~y2)
children(Model(e), ~x2)
children(Model(e), ~x2+eta)
effects(e,y2~x1)
## All simple paths (undirected)
path(m,y1~x1,all=TRUE)

[Package lava version 1.6.5 Index]